+++ packages/dumb-durable-websocket/index.ts
@@ -1,15 +1,15 @@
-export type WebSocketServer = {
+export type WebSocketPool = {
   sessions: WebSocket[];
 };
 
-export const ws = (): WebSocketServer => {
+export const ws = (): WebSocketPool => {
   return {
     sessions: [],
   };
 };
 
-export const disconnectAll = (
-  ws: WebSocketServer,
+export const close = (
+  ws: WebSocketPool,
   { code, reason }: Partial<WebSocketDisconnect> = {}
 ) => {
   const closing = ws.sessions;
@@ -21,40 +21,22 @@ export const disconnectAll = (
 };
 
 export const connect = (
-  ws: WebSocketServer,
-  request: Request,
-  { onAccept, onMessage, onOpen, onClose, onError }: ConnectHandlerConfig
+  ws: WebSocketPool,
+  setup: (socket: WebSocket) => void | Promise<void>
 ): Response => {
-
-
   const pair = new WebSocketPair();
   const websocket = pair[1];
 
   websocket.accept();
   ws.sessions.push(websocket);
-  onAccept?.(pair[1]);
 
-  const handleOnMessage = (ev: MessageEvent) => onMessage?.(pair[1], ev);
-  const handleOnOpen = (ev: Event) => onOpen?.(pair[1], ev);
-  const handleOnError = (ev: ErrorEvent) => onError?.(pair[1], ev);
-  const handleOnClose = (ev: CloseEvent) => {
-    websocket.removeEventListener("message", handleOnMessage);
-    websocket.removeEventListener("open", handleOnOpen);
-    websocket.removeEventListener("error", handleOnError);
-    websocket.removeEventListener("close", handleOnClose);
-    onClose?.(pair[1], ev);
-  };
-
-  websocket.addEventListener("message", handleOnMessage);
-  websocket.addEventListener("open", handleOnOpen);
-  websocket.addEventListener("error", handleOnError);
-  websocket.addEventListener("close", handleOnClose);
+  setup(pair[1]);
 
   return new Response(null, { status: 101, webSocket: pair[0] });
 };
 
 export const disconnect = (
-  ws: WebSocketServer,
+  ws: WebSocketPool,
   websocket: WebSocket,
   { code, reason }: Partial<WebSocketDisconnect> = {}
 ) => {
@@ -63,7 +45,7 @@ export const disconnect = (
 };
 
 export const broadcast = (
-  ws: WebSocketServer,
+  ws: WebSocketPool,
   message: string,
   options: { exclude: WebSocket[] } = { exclude: [] }
 ) => {
@@ -81,18 +63,11 @@ export const broadcast = (
   });
 };
 
-export const isNotWebSocketUpgrade = (request: Request) => {
-  return request.headers.get("Upgrade") !== "websocket";
-};
-
-export type SocketController = ReturnType<typeof ws>;
-
-export type ConnectHandlerConfig = {
-  onAccept?: (websocket: WebSocket) => Promise<void> | void;
-  onMessage?: (websocket: WebSocket, ev: MessageEvent) => Promise<void> | void;
-  onError?: (websocket: WebSocket, ev: ErrorEvent) => Promise<void> | void;
-  onClose?: (websocket: WebSocket, ev: CloseEvent) => Promise<void> | void;
-  onOpen?: (websocket: WebSocket, ev: Event) => Promise<void> | void;
+export const isWebSocketUpgrade = (request: Request) => {
+  return (
+    request.headers.get("Upgrade")?.toLowerCase() === "websocket" &&
+    request.method.toLowerCase() === "get"
+  );
 };
 
 export type WebSocketDisconnect = { code: number; reason: string };